+++ extras/PackageMaintenance.R
@@ -35,12 +35,13 @@ shell("R CMD Rd2pdf ./ --output=extras/Capr.pdf")
 dir.create(path = "./extras/pdf_vignette/", showWarnings = FALSE)
 rmarkdown::render("vignettes/capr_design.Rmd",
                   output_file = "../extras/pdf_vignette/capr_design.pdf",
+
   rmarkdown::pdf_document(latex_engine = "pdflatex", toc = TRUE, number_sections = TRUE))
 unlink("extras/pdf_vignette/capr_design.tex")
 
 
 # Examples
-dir.create(path = "./extras/pdf_vignette/", showWarnings = FALSE)
+#dir.create(path = "./extras/pdf_vignette/", showWarnings = FALSE)
 rmarkdown::render("vignettes/Examples.Rmd",
                   output_file = "../extras/pdf_vignette/Examples.pdf",
                   rmarkdown::pdf_document(latex_engine = "pdflatex",
@@ -49,25 +50,34 @@ unlink("extras/pdf_vignette/Examples.tex")
 
 
 # Using Capr
-dir.create(path = "./extras/pdf_vignette/", showWarnings = FALSE)
+#dir.create(path = "./extras/pdf_vignette/", showWarnings = FALSE)
 rmarkdown::render("vignettes/Using-Capr.Rmd", output_file = "../extras/pdf_vignette/Using-Capr.pdf",
   rmarkdown::pdf_document(latex_engine = "pdflatex", toc = TRUE, number_sections = TRUE))
 unlink("extras/pdf_vignette/Using-Capr.tex")
 
 
-# capr templates
-dir.create(path = "./extras/pdf_vignette/", showWarnings = FALSE)
-rmarkdown::render("vignettes/capr_templates.Rmd", output_file = "../extras/pdf_vignette/capr_templates.pdf",
-                  rmarkdown::pdf_document(latex_engine = "pdflatex", toc = TRUE, number_sections = TRUE))
+# capr templates dir.create(path = './extras/pdf_vignette/', showWarnings = FALSE)
+rmarkdown::render("vignettes/capr_templates.Rmd",
+                  output_file = "../extras/pdf_vignette/capr_templates.pdf",
+
+  rmarkdown::pdf_document(latex_engine = "pdflatex", toc = TRUE, number_sections = TRUE))
 unlink("extras/pdf_vignette/capr_templates.tex")
 
 
-# capr concept sets
-dir.create(path = "./extras/pdf_vignette/", showWarnings = FALSE)
-rmarkdown::render("vignettes/Capr-conceptSets.Rmd", output_file = "../extras/pdf_vignette/Capr-conceptSets.pdf",
-                  rmarkdown::pdf_document(latex_engine = "pdflatex", toc = TRUE, number_sections = TRUE))
+# capr concept sets dir.create(path = './extras/pdf_vignette/', showWarnings = FALSE)
+rmarkdown::render("vignettes/Capr-conceptSets.Rmd",
+                  output_file = "../extras/pdf_vignette/Capr-conceptSets.pdf",
+
+  rmarkdown::pdf_document(latex_engine = "pdflatex", toc = TRUE, number_sections = TRUE))
 unlink("extras/pdf_vignette/Capr-conceptSets.tex")
 
+
+# capr components
+rmarkdown::render("vignettes/capr_objects.Rmd",
+                  output_file = "../extras/pdf_vignette/capr_objects.pdf",
+
+  rmarkdown::pdf_document(latex_engine = "pdflatex", toc = TRUE, number_sections = TRUE))
+unlink("extras/pdf_vignette/capr_objects.tex")
 
 # build site
 pkgdown::build_site()