+++ tests/testthat/test-collectCodesetId.R
@@ -4,19 +4,19 @@
 # Query, group
 
 test_that("listConceptSets - Query", {
-  conceptSets <- listConceptSets(condition(cs(1)))
+  conceptSets <- listConceptSets(conditionOccurrence(cs(1, name = "test")))
   expect_true(all(purrr::map_lgl(conceptSets, ~all(names(.) == c("id", "name", "expression")))))
 })
 
 test_that("listConceptSets - Criteria", {
-  conceptSets <- listConceptSets(atLeast(1, condition(cs(1))))
+  conceptSets <- listConceptSets(atLeast(1, conditionOccurrence(cs(1, name = "test"))))
   expect_true(all(purrr::map_lgl(conceptSets, ~all(names(.) == c("id", "name", "expression")))))
 })
 
 test_that("listConceptSets - Group", {
   g <- withAll(
-    atLeast(1, condition(cs(1))),
-    exactly(0, condition(cs(2)))
+    atLeast(1, conditionOccurrence(cs(1, name = "test"))),
+    exactly(0, conditionOccurrence(cs(2, name = "test")))
   )
   conceptSets <- listConceptSets(g)
   expect_length(conceptSets, 2)
@@ -25,8 +25,8 @@ test_that("listConceptSets - Group", {
 
 test_that("listConceptSets - Entry", {
   e <- entry(withAll(
-    atLeast(1, condition(cs(1))),
-    exactly(0, condition(cs(2)))
+    atLeast(1, conditionOccurrence(cs(1, name = "test"))),
+    exactly(0, conditionOccurrence(cs(2, name = "test")))
   ))
   conceptSets <- listConceptSets(e)
   expect_true(all(purrr::map_lgl(conceptSets, ~all(names(.) == c("id", "name", "expression")))))
@@ -34,8 +34,8 @@ test_that("listConceptSets - Entry", {
 
 test_that("listConceptSets - Attrition", {
   x <- attrition(withAll(
-    atLeast(1, condition(cs(1))),
-    exactly(0, condition(cs(2)))
+    atLeast(1, conditionOccurrence(cs(1, name = "test"))),
+    exactly(0, conditionOccurrence(cs(2, name = "test")))
   ))
 
   conceptSets <- listConceptSets(x)
@@ -48,14 +48,14 @@ test_that("listConceptSets - Attrition", {
   x <- attrition(
     'no t1d' = withAll(
       exactly(0,
-              condition(cs(descendants(201254L))),
+              conditionOccurrence(cs(descendants(201254L), name = "test")),
               duringInterval(eventStarts(-Inf, -1))
       )
     ),
     'abnormal hba1c' = withAll(
       atLeast(1,
               measurement(
-                cs(descendants(4184637L)),
+                cs(descendants(4184637L), name = "test"),
                 valueAsNumber(lt(13)),
                 unit(8713L)
               ),
@@ -73,7 +73,7 @@ test_that("listConceptSets - Attrition", {
 
 
 test_that("listConceptSets - CohortExit", {
-  e <- exit(drugExit(cs(1,2,5)))
+  e <- exit(drugExit(cs(1,2,5, name = "test")))
   expect_s4_class(e, "CohortExit")
   conceptSets <- listConceptSets(e)
   expect_length(conceptSets, 1)
@@ -81,9 +81,9 @@ test_that("listConceptSets - CohortExit", {
 })
 
 test_that("listConceptSets - Query with nested criteria", {
-  x <- visit(cs(1),
-             nestedWithAll(atLeast(1, condition(cs(9))),
-                           atLeast(1, drug(cs(1:5)))))
+  x <- visit(cs(1, name = "test"),
+             nestedWithAll(atLeast(1, conditionOccurrence(cs(9, name = "test"))),
+                           atLeast(1, drugExposure(cs(1:5, name = "test")))))
 
   conceptSets <- listConceptSets(x)
   expect_length(conceptSets, 3)
@@ -91,11 +91,11 @@ test_that("listConceptSets - Query with nested criteria", {
 })
 
 test_that("listConceptSets - Query with double nested criteria", {
-  x <- visit(cs(1),
+  x <- visit(cs(1, name = "test"),
          nestedWithAll(
-           atLeast(1, drug(cs(21:24))),
-           atLeast(1, condition(cs(9),
-             nestedWithAll(atLeast(1, drug(cs(11))))
+           atLeast(1, drugExposure(cs(21:24, name = "test"))),
+           atLeast(1, conditionOccurrence(cs(9, name = "test"),
+             nestedWithAll(atLeast(1, drugExposure(cs(11, name = "test"))))
       ))))
 
   conceptSets <- listConceptSets(x)
@@ -107,13 +107,13 @@ test_that("listConceptSets - nested Query", {
   skip("failing test") # TODO fix listConceptSets so this passes
 
   # this works fine
-  x <- visit(cs(descendants(9201, 9203, 262)),
+  x <- visit(cs(descendants(9201, 9203, 262), name = "test"),
     nestedWithAll(
       atLeast(1,
-        condition(cs(descendants(316139), name = "heart failure"),
+        conditionOccurrence(cs(descendants(316139), name = "heart failure"),
           attributes = nestedWithAll(
             atLeast(1,
-              condition(cs(descendants(316139), name = "heart failure"))
+              conditionOccurrence(cs(descendants(316139), name = "heart failure"))
             )
           )
         )
@@ -127,13 +127,13 @@ test_that("listConceptSets - nested Query", {
   expect_true(all(purrr::map_lgl(conceptSets, ~all(names(.) == c("id", "name", "expression")))))
 
   # this does not work
-  x <- visit(cs(descendants(9201, 9203, 262)),
+  x <- visit(cs(descendants(9201, 9203, 262), name = "test"),
     nestedWithAll(
       atLeast(1,
-        condition(cs(descendants(316139), name = "heart failure"),
+        conditionOccurrence(cs(descendants(316139), name = "heart failure"),
           attributes = list(male(), nestedWithAll(
             atLeast(1,
-              condition(cs(descendants(316139), name = "heart failure"))
+              conditionOccurrence(cs(descendants(316139), name = "heart failure"))
             )
           ))
         )
@@ -151,20 +151,20 @@ test_that("listConceptSets - Cohort", {
 
   cd <- cohort(
     entry = entry(
-      condition(cs(descendants(201826L)), male()),
+      conditionOccurrence(cs(descendants(201826L), name = "test"), male()),
       observationWindow = continuousObservation(365, 0)
     ),
     attrition = attrition(
       'no t1d' = withAll(
         exactly(0,
-                condition(cs(descendants(201254L))),
+                conditionOccurrence(cs(descendants(201254L), name = "test")),
                 duringInterval(eventStarts(-Inf, -1))
         )
       ),
       'abnormal hba1c' = withAll(
         atLeast(1,
                 measurement(
-                  cs(descendants(4184637L)),
+                  cs(descendants(4184637L), name = "test"),
                   valueAsNumber(lt(13)),
                   unit(8713L)
                 ),